+++ bestmobabot/vk.py
@@ -2,33 +2,29 @@
 VK.com API wrapper.
 """
 
-import contextlib
 import re
 from typing import Iterable
 
-import requests
 from loguru import logger
+from requests import Session
 
 from bestmobabot import constants
 from bestmobabot.settings import Settings
 
 
-class VK(contextlib.AbstractContextManager):
+class VK:
     URL = 'https://api.vk.com/method/wall.get'
     GIFT_ID_RE = re.compile(r'gift_id=(\w+)')
     VK_CC_RE = re.compile(r'https://vk.cc/\w+')
 
-    def __init__(self, settings: Settings):
-        self.session = requests.Session()
+    def __init__(self, session: Session, settings: Settings):
+        self.session = session
         self.params = {
             'access_token': settings.vk.access_token,
             'owner_id': '-116039030',
             'count': '5',
             'v': '5.92',
         }
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.session.__exit__(exc_type, exc_val, exc_tb)
 
     def find_gifts(self) -> Iterable[str]:
         logger.info('Checking VK.com gifts…')