+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add dungeon APIs
 - Add `dungeon_activity` reward
 - Improve timezone logging
+- Use single `requests.Session`
 
 ## `3.0b2`